backend/requirements/requirements.txt
alembic==1.8.0
anyio==3.6.1
asgiref==3.5.2
asyncpg==0.26.0
attrs==21.4.0
certifi==2022.6.15
cfgv==3.3.1
charset-normalizer==2.1.0
click==8.1.3
distlib==0.3.4
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.78.0
filelock==3.7.1
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
identify==2.5.1
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.6.0
Mako==1.2.1
MarkupSafe==2.1.1
nodeenv==1.7.0
orjson==3.7.7
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pydantic==1.9.1
python-decouple==3.6
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sqladmin==0.1.12
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.10
uvicorn==0.17.6
uvloop==0.16.0
virtualenv==20.15.1
watchgod==0.8.2
websockets==10.3
wtforms==3.0.1